+++ scripts/stop_clamav.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
 
 sudo systemctl stop clamav-freshclam
+sudo systemctl stop clamav-daemon.socket
 sudo systemctl stop clamav-daemon 
 
 sudo systemctl disable clamav-freshclam
 sudo systemctl disable clamav-daemon.socket
-
+sudo systemctl disable clamav-daemon